+++ ExposurePlayback/Context/Components/HLSNative+ExposureContext+Playback.swift
@@ -82,7 +82,11 @@ extension Player where Tech == HLSNative<ExposureContext> {
             source.analyticsConnector.providers.forEach{
                 if let exposureProvider = $0 as? ExposureStreamingAnalyticsProvider {
                     exposureProvider.onHandshakeStarted(tech: tech, source: source)
-                    exposureProvider.finalizePreparation(tech: tech, source: source, playSessionId: source.entitlement.playSessionId, heartbeatsProvider: tech)
+                    exposureProvider.finalizePreparation(tech: tech, source: source, playSessionId: source.entitlement.playSessionId) { [weak self] in
+                        guard let `self` = self else { return nil }
+                        guard let heartbeatsProvider = source as? HeartbeatsProvider else { return nil }
+                        return heartbeatsProvider.heartbeat(for: self.tech, in: self.context)
+                    }
                 }
             }
         }