+++ ExposurePlayback/Context/Source/ChannelSource.swift
@@ -107,3 +107,14 @@ extension ChannelSource: ProgramServiceEnabled {
         return assetId
     }
 }
+
+extension ChannelSource: HeartbeatsProvider {
+    internal func heartbeat(for tech: HLSNative<ExposureContext>, in context: ExposureContext) -> Playback.Heartbeat {
+        if isUnifiedPackager {
+            return Playback.Heartbeat(timestamp: Date().millisecondsSince1970, offsetTime: tech.playheadTime ?? tech.playheadPosition)
+        }
+        else {
+            return Playback.Heartbeat(timestamp: Date().millisecondsSince1970, offsetTime: tech.playheadPosition)
+        }
+    }
+}