+++ ExposurePlayback/Protocols/ExposureStreamingAnalyticsProvider.swift
@@ -45,7 +45,7 @@ public protocol ExposureStreamingAnalyticsProvider: AnalyticsProvider {
     /// - parameter source: `MediaSource` used to load the request
     /// - parameter playSessionId: Unique identifier for the current playback session.
     /// - parameter heartbeatsProvider: Will deliver heartbeats metadata during the session
-    func finalizePreparation<Tech, Source>(tech: Tech, source: Source, playSessionId: String, heartbeatsProvider: HeartbeatsProvider) where Tech: PlaybackTech, Source: MediaSource
+    func finalizePreparation<Tech, Source>(tech: Tech, source: Source, playSessionId: String, heartbeatsProvider: @escaping () -> AnalyticsEvent?) where Tech: PlaybackTech, Source: MediaSource
     
     /// Sent if the current program changes during the session.
     ///