+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # CHANGELOG
 
+* `2.0.78` Release - [2.0.78](#2078)
 * `0.77.x` Releases - [0.77.0](#0770)
+
+## 2.0.78
+
+#### Changes
+* Standalone networking
+* `HeartbeatsProvider` as a closure instead of a protocol
+
+#### Bugfixes
+* Heartbeats report offset based on stream type
 
 ## 0.77.0